cloudconfig: refactor config
- Move CloudConfig into config package
- Add YAML tags to CloudConfig
